requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
whitenoise
psycopg2-binary
dj-database-url
